refactor: centralize password validation rules in AppServiceProvider
In PR #10 (https://github.com/LucaPipolo/budget-app/pull/10), we set up
custom password validation rules by overriding the default Filament
Register page and changing the Laravel Fortify settings.

It turned out that a better way to achieve this would be to set default
password validation rules in the `AppServiceProvider`. This will also
cover cases that were not covered with the previous implementation, such
as the Filament password recovery form.